// src/app/models/IMotivo.ts
export interface IMotivo {
  id?: number;
  tiempoXMantenimiento: number;
  cambiosIngenieria: number;
  cambioModelo: number;
  ausentismo: number;
  dobladoras: number;
  soldadura: number;
  montaje: number;
  IAPlacasDisplay: number;
  IAPlacasMain: number;
  IMPlacasDisplay: number;
  IMPlacasMain: number;
  equipamientoMaquinariaMant: number;
  equipamientoMaquinariaIng: number;
  metodos: number;
  calidad: number;
  sistemas: number;
  it: number;
  abastecimiento: number;
  cli: number;
  supply: number;
  terceros: number;
  otros: number;
}